requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
torch
torchvision
scikit-learn
scipy
matplotlib
scikit-image
tqdm
    